Terraform/cloud-fun/requirements.txt
functions-framework==3.*
requests
google-cloud-pubsub